test(notebook-tools,#14615): consolidate check_twin_parity duplicate suite (family 6/6, part 2)
Port the EPIC #12933 NUMBERING-DRIFT layer (16 tests + git mini-repo
fixtures) from scripts/tests/test_check_twin_parity.py into the canonical
notebook_tools/tests suite, collocated with the module. Import style
normalized (direct imports instead of per-test importlib reload - the
module has no mutable global state and the canonical suite already
imports it at top level).

Deduplicated: the 3 content-drift combos of the legacy
test_classify_per_pair ((OK,DRIFT), (DRIFT,OK), (OK,OK)) are already
pinned by the canonical 9-combo EXPECTED table (c.709); only the 4
NUMBERING-DRIFT combos are ported (renamed
test_classify_numbering_per_pair).

Count proof: canonical 13 items + 30 ported = 43 passed (legacy file
collected 33 items; 33 - 3 dedupes = 30). scripts/tests collection
4631 -> 4598 = exactly the 33 removed.
